Moved the door site connector to a plugin
