+++ door/plugins/garage_door_site/index.js
@@ -1,0 +1,14 @@
+var Site = require('./lib/site')
+
+var register = function(server, options, next) {
+  site = new Site(options)
+  server.expose('site', site)
+
+  return next()
+}
+
+register.attributes = {
+  pkg: require('./package.json')
+}
+
+module.exports = register
