+++ door/plugins/garage_door_site/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "garage_door_site",
+  "version": "0.1.0"
+}
